add lifecycle context + add api to context
